remove mark synchronous
